+++ src/extension.ts
@@ -6,11 +6,17 @@ export function activate(context: vscode.ExtensionContext) {
 	registerExcludeWatcher(context);
 
 	context.subscriptions.push(
-		vscode.commands.registerCommand('git-ignore-manager.addToGitignore', async (resource?: vscode.Uri) => {
-			await addResourceToIgnore(resource, 'gitignore');
+		vscode.commands.registerCommand('git-ignore-manager.addToGitignore', async (
+			resource?: vscode.Uri,
+			selectedResources?: vscode.Uri[],
+		) => {
+			await addResourceToIgnore(resource, selectedResources, 'gitignore');
 		}),
-		vscode.commands.registerCommand('git-ignore-manager.addToExclude', async (resource?: vscode.Uri) => {
-			await addResourceToIgnore(resource, 'exclude');
+		vscode.commands.registerCommand('git-ignore-manager.addToExclude', async (
+			resource?: vscode.Uri,
+			selectedResources?: vscode.Uri[],
+		) => {
+			await addResourceToIgnore(resource, selectedResources, 'exclude');
 		}),
 		vscode.commands.registerCommand('git-ignore-manager.openGitignore', async (resource?: vscode.Uri) => {
 			await openIgnoreFile(resource, 'gitignore');